Anonymous wird nicht mehr angezeigt
Entfernen von Anonymous aus der Benutzerliste im Management

+++ cms/management/benutzer/index.php
@@ -43,31 +43,33 @@
 					}
 					echo '<th colspan="2">Benutzer &#228;ndern</th></tr>';
 					foreach($alleBenutzer as $currBenutzer) {
-						$BenutzerRights = $myADBConnector->getOneBenutzer($currBenutzer['UID']);
-						echo '<tr><td>'.$currBenutzer['Uname'].'</td>';
-						$i = 0;
-						foreach($alleRollen as $currentRolle) {
-							if($currBenutzer['UID'] <= 2) {
-								if($BenutzerRights[$i]['Xvalue'] == 1)
-									echo '<td id="important_green">YES</td>';
-								else
-									echo '<td id="important_red">NO</td>';
+						if($currBenutzer['UID'] != 2) {
+							$BenutzerRights = $myADBConnector->getOneBenutzer($currBenutzer['UID']);
+							echo '<tr><td>'.$currBenutzer['Uname'].'</td>';
+							$i = 0;
+							foreach($alleRollen as $currentRolle) {
+								if($currBenutzer['UID'] <= 2) {
+									if($BenutzerRights[$i]['Xvalue'] == 1)
+										echo '<td id="important_green">YES</td>';
+									else
+										echo '<td id="important_red">NO</td>';
+								}
+								else {
+									if($BenutzerRights[$i]['Xvalue'] == 1)
+										echo '<td id="important_green"><a href="save.php?UID='.$currBenutzer['UID'].'&RID='.($i+1).'&Xvalue=0">YES</a></td>';
+									else
+										echo '<td id="important_red"><a href="save.php?UID='.$currBenutzer['UID'].'&RID='.($i+1).'&Xvalue=1">NO</a></td>';
+								}
+								$i++;
 							}
-							else {
-								if($BenutzerRights[$i]['Xvalue'] == 1)
-									echo '<td id="important_green"><a href="save.php?UID='.$currBenutzer['UID'].'&RID='.($i+1).'&Xvalue=0">YES</a></td>';
-								else
-									echo '<td id="important_red"><a href="save.php?UID='.$currBenutzer['UID'].'&RID='.($i+1).'&Xvalue=1">NO</a></td>';
+							if($currBenutzer['UID'] > 2) {
+								echo '<td id="important_green"><a href="PWmask.php?UID='.$currBenutzer['UID'].'">Passwort &#228;ndern</a></td>';
+								echo '<td id="important_red"><a href="delete.php?UID='.$currBenutzer['UID'].'">l&#246;schen</a></td>';
 							}
-							$i++;
-						}
-						if($currBenutzer['UID'] > 2) {
-							echo '<td id="important_green"><a href="PWmask.php?UID='.$currBenutzer['UID'].'">Passwort &#228;ndern</a></td>';
-							echo '<td id="important_red"><a href="delete.php?UID='.$currBenutzer['UID'].'">l&#246;schen</a></td>';
-						}
-						else {
-							echo '<td id="important_green"><a href="PWmask.php?UID='.$currBenutzer['UID'].'">Passwort &#228;ndern</a></td>';
-							echo '<td>l&#246;schen</td></tr>';
+							if($currBenutzer['UID'] == 1) {
+								echo '<td id="important_green"><a href="PWmask.php?UID='.$currBenutzer['UID'].'">Passwort &#228;ndern</a></td>';
+								echo '<td>l&#246;schen</td></tr>';
+							}
 						}
 					}					
 					echo '</table>';
